LPS-170489 Inline ModuleAggregareClassLoader to LiferayServiceExtender, since only LiferayServiceExtender invokes it for now.
